+++ priv/repo/seeds.exs
@@ -11,10 +11,6 @@
 # and so on) as they will fail if something goes wrong.
 
 alias Siwapp.{
-  Commons,
-  Customers,
-  Invoices,
-  RecurringInvoices,
   Settings,
   Templates
 }
@@ -52,113 +48,3 @@ Templates.create(%{
   })
 
 Templates.set_default(:email, email_template)
-
-# SEEDING SERIES
-series = [
-  %{name: "A-series", code: "A"},
-  %{name: "B-series", code: "B"},
-  %{name: "C-series", code: "C"}
-]
-
-Enum.each(series, &Commons.create_series(&1))
-
-# SEEDING TAXES
-taxes = [
-  %{name: "VAT", value: 21, default: true},
-  %{name: "RETENTION", value: -15}
-]
-
-Enum.each(taxes, &Commons.create_tax(&1))
-
-# SEEDING CUSTOMERS
-customers = Enum.map(0..15, fn _i -> %{name: Faker.Person.name(), id: Faker.Code.issn()} end)
-
-Enum.each(
-  customers,
-  &Customers.create(%{
-    name: &1.name,
-    identification: &1.id,
-    email: Faker.Internet.email(),
-    contact_person: Faker.Person.name(),
-    invoicing_address:
-      "#{Faker.Address.street_address()}\n#{Faker.Address.postcode()} #{Faker.Address.country()}"
-  })
-)
-
-# SEEDING INVOICES
-currencies = ["USD", "USD", "USD", "EUR", "GBP"]
-booleans = [true, false]
-
-invoices =
-  Enum.map(0..30, fn _i ->
-    %{customer: Enum.random(customers), issue_date: Faker.Date.backward(31)}
-  end)
-
-Enum.each(
-  invoices,
-  &Invoices.create(%{
-    name: &1.customer.name,
-    identification: &1.customer.id,
-    paid: Enum.random(booleans),
-    sent_by_email: Enum.random(booleans),
-    issue_date: &1.issue_date,
-    due_date: Date.add(&1.issue_date, Faker.random_between(1, 31)),
-    series_id: Faker.random_between(1, 3),
-    currency: Enum.random(currencies),
-    items: [
-      %{
-        quantity: Faker.random_between(1, 2),
-        description: "#{Faker.App.name()} App Development",
-        unitary_cost: Faker.random_between(10_000, 1_000_000),
-        taxes: ["VAT", "RETENTION"]
-      }
-    ]
-  })
-)
-
-# SEEDING RECURRING INVOICES
-today = Date.utc_today()
-
-recurring_invoices = [
-  %{
-    name: Faker.Person.name(),
-    period: 3,
-    period_type: "Daily",
-    starting_date: Date.add(today, -30),
-    series_id: 1
-  },
-  %{
-    name: Faker.Person.name(),
-    period: 2,
-    period_type: "Monthly",
-    starting_date: Date.add(today, -60),
-    series_id: 2,
-    items: %{
-      "0" => %{
-        "quantity" => Faker.random_between(1, 2),
-        "description" => "#{Faker.App.name()} App Development",
-        "unitary_cost" => Faker.random_between(10_000, 1_000_000),
-        "discount" => 0,
-        "taxes" => ["RETENTION"]
-      }
-    }
-  },
-  %{
-    name: Faker.Person.name(),
-    period: 1,
-    period_type: "Yearly",
-    starting_date: Date.add(today, -400),
-    series_id: 3,
-    items: %{
-      "0" => %{
-        "quantity" => Faker.random_between(1, 2),
-        "description" => "#{Faker.App.name()} App Development",
-        "unitary_cost" => Faker.random_between(10_000, 1_000_000),
-        "discount" => 0,
-        "taxes" => ["VAT"]
-      }
-    }
-  }
-]
-
-Enum.each(recurring_invoices, &RecurringInvoices.create(&1))
